registry error fix
