+++ src/app.ts
@@ -4,14 +4,17 @@ import cors from "cors";
 import swaggerUi from "swagger-ui-express";
 import { openApiDoc } from "./utils/swagger";
 import { errorHandler } from "./middlewares/errorHandler";
-import authRoutes from "./modules/auth/routes/authRoutes";
+import userRoutes from "./modules/User/userRoutes";
+import authRoutes from "./modules/auth/authRoutes";
 import protectedRoutes from "./modules/protected/routes/protectedRoutes";
 import donorRoutes from "./modules/donor/donorRoutes";
 import AccountRoutes from "./modules/Account/AccountRoutes";
-import systemAdminRoutes  from "./modules/systemAdmin/SystemAdminRoutes";
-import HospitalRoutes  from "./modules/Hospital/HospitalRoutes";
-import BloodRoutes  from "./modules/Blood/BloodRoutes";
+import systemAdminRoutes from "./modules/systemAdmin/SystemAdminRoutes";
+import HospitalRoutes from "./modules/Hospital/HospitalRoutes";
+import BloodRoutes from "./modules/Blood/BloodRoutes";
 import inventoryRoutes from "./modules/Inventory/InventoryRoutes";
+import { extendZodWithOpenApi } from "@asteasolutions/zod-to-openapi/dist/zod-extensions";
+import { z } from "zod";
 
 dotenv.config();
 
@@ -29,14 +32,15 @@ app.use(cors(corsOptions));
 app.use(errorHandler);
 
 // Routes
+app.use("/api", userRoutes);
 app.use("/auth", authRoutes);
 app.use("/api", protectedRoutes);
 app.use("/api", donorRoutes);
 app.use("/api", AccountRoutes);
-app.use("/api",systemAdminRoutes);
-app.use("/api",HospitalRoutes);
-app.use("/api",BloodRoutes);
-app.use("/api",inventoryRoutes);
+app.use("/api", systemAdminRoutes);
+app.use("/api", HospitalRoutes);
+app.use("/api", BloodRoutes);
+app.use("/api", inventoryRoutes);
 app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(openApiDoc));
 
 export default app;
