+++ src/modules/User/userRoutes.ts
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { register } from "./userController";
+
+const router = Router();
+
+router.post("user/register", register);
+
+export default router;
