+++ src/modules/auth/authController.ts
@@ -1,0 +1,26 @@
+import { Request, Response } from "express";
+
+import { createUser } from "../User/userService";
+import { validateAndParse } from "../../utils/validateAndParseRequest";
+import { NewUserSchema } from "../User/userSchema";
+import { authenticateUser, refreshAccessToken } from "./authService";
+
+export const login = async (req: Request, res: Response) => {
+  const { email, password } = req.body;
+  const tokens = await authenticateUser(email, password);
+  if (tokens) {
+    res.json(tokens);
+  } else {
+    res.status(401).json({ error: "Invalid email or password" });
+  }
+};
+
+export const refreshToken = (req: Request, res: Response) => {
+  const { refreshToken } = req.body;
+  const newAccessToken = refreshAccessToken(refreshToken);
+  if (newAccessToken) {
+    res.json({ accessToken: newAccessToken });
+  } else {
+    res.status(401).json({ error: "Invalid refresh token" });
+  }
+};
