+++ src/modules/auth/authRoutes.ts
@@ -1,10 +1,9 @@
 import { Router } from "express";
-import { login, register, refreshToken } from "../controllers/authController";
+import { login, refreshToken } from "./authController";
 
 const router = Router();
 
 router.post("/login", login);
-router.post("/register", register);
 router.post("/refresh-token", refreshToken);
 
 export default router;
